Refactor interpreters + Change Fetch.traverse
- Create a interpreter package and split the interpreter parts in three
  separate files.
- Fix a bug in the handling of `FetchMany`
- Customize the implementation of `Fetch.traverse` to reduce the stack
  needed to simplify/optimize the resulting `Fetch`
