+++ monix/shared/src/test/scala/FetchTaskTests.scala
@@ -41,9 +41,8 @@ class FetchTaskTests extends AsyncFreeSpec with Matchers {
       Query.async((ok, fail) => {
         ok(Option(Article(id.id, "An article with id " + id.id)))
       })
-    override def fetchMany(ids: NonEmptyList[ArticleId]): Query[Map[ArticleId, Article]] = {
+    override def fetchMany(ids: NonEmptyList[ArticleId]): Query[Map[ArticleId, Article]] =
       batchingNotSupported(ids)
-    }
   }
 
   def article(id: Int): Fetch[Article] = Fetch(ArticleId(id))
@@ -57,9 +56,8 @@ class FetchTaskTests extends AsyncFreeSpec with Matchers {
       Query.async((ok, fail) => {
         ok(Option(Author(id.id, "@egg" + id.id)))
       })
-    override def fetchMany(ids: NonEmptyList[AuthorId]): Query[Map[AuthorId, Author]] = {
+    override def fetchMany(ids: NonEmptyList[AuthorId]): Query[Map[AuthorId, Author]] =
       batchingNotSupported(ids)
-    }
   }
 
   def author(a: Article): Fetch[Author] = Fetch(AuthorId(a.author))
