+++ shared/src/main/scala/datasource.scala
@@ -22,34 +22,34 @@ import cats.syntax.functor._
 import cats.syntax.traverseFilter._
 
 /**
-  * A `DataSource` is the recipe for fetching a certain identity `I`, which yields
-  * results of type `A`.
-  */
+ * A `DataSource` is the recipe for fetching a certain identity `I`, which yields
+ * results of type `A`.
+ */
 trait DataSource[I, A] {
 
   /** The name of the data source.
-    */
+   */
   def name: DataSourceName
   override def toString: String = name
 
   /**
-    * Derive a `DataSourceIdentity` from an identity, suitable for storing the result
-    * of such identity in a `DataSourceCache`.
-    */
+   * Derive a `DataSourceIdentity` from an identity, suitable for storing the result
+   * of such identity in a `DataSourceCache`.
+   */
   def identity(i: I): DataSourceIdentity = (name, i)
 
   /** Fetch one identity, returning a None if it wasn't found.
-    */
+   */
   def fetchOne(id: I): Query[Option[A]]
 
   /** Fetch many identities, returning a mapping from identities to results. If an
-    * identity wasn't found won't appear in the keys.
-    */
+   * identity wasn't found won't appear in the keys.
+   */
   def fetchMany(ids: NonEmptyList[I]): Query[Map[I, A]]
 
   /** Use `fetchOne` for implementing of `fetchMany`. Use only when the data
-    * source doesn't support batching.
-    */
+   * source doesn't support batching.
+   */
   def batchingNotSupported(ids: NonEmptyList[I]): Query[Map[I, A]] = {
     val fetchOneWithId: I => Query[Option[(I, A)]] = id =>
       OptionT(fetchOne(id)).map(res => (id, res)).value
@@ -66,5 +66,5 @@ trait DataSource[I, A] {
 }
 
 sealed trait ExecutionType extends Product with Serializable
-case object Sequential extends ExecutionType
-case object Parallel extends ExecutionType
+case object Sequential     extends ExecutionType
+case object Parallel       extends ExecutionType
